+++ safe-coloured-text-parsing-gen/src/Text/Colour/Chunk/Parsing/Gen.hs
@@ -6,6 +6,10 @@ import Data.GenValidity
 import Text.Colour.Chunk.Parsing
 import Text.Colour.Gen ()
 
+instance GenValid OscCommand where
+  genValid = genValidStructurallyWithoutExtraChecking
+  shrinkValid = shrinkValidStructurallyWithoutExtraFiltering
+
 instance GenValid AnsiToken where
   genValid = genValidStructurallyWithoutExtraChecking
   shrinkValid = shrinkValidStructurallyWithoutExtraFiltering